requirements.txt
fastapi
uvicorn
pydantic
selenium>=4.36.0
webdriver-manager>=4.0.2
undetected-chromedriver>=3.5.5
setuptools>=65.0.0
python-dotenv==1.0.0
2captcha-python==1.1.3
requests==2.31.0
numpy>=2.0.0
python-multipart>=0.0.9
